+++ NEWS.md
@@ -1,3 +1,6 @@
+
+- `read_resource()` will now warn rather than error on unknown encoding (#86).
+
 # frictionless 0.10.0
 
 - `add_resource()` now supports adding CSV file(s) directly as a resource.